Bug 1674777 part 5 - Change GetObjectAs* APIs to use size_t for outparam instead of uint32_t. r=sfink,sg
Note that in the IndexedDB code we pass a Span to Key::EncodeAsString where we check
for large lengths.

Differential Revision: https://phabricator.services.mozilla.com/D103679

UltraBlame original commit: b142ae1ffe745c57aac35de2fcb69bfe94f43a48

+++ dom/indexedDB/Key.cpp
@@ -834,7 +834,7 @@ double Key::DecodeNumber(const EncodedDataType*& aPos,
 Result<Ok, nsresult> Key::EncodeBinary(JSObject* aObject, bool aIsViewObject,
                                        uint8_t aTypeOffset) {
   uint8_t* bufferData;
-  uint32_t bufferLength;
+  size_t bufferLength;
